ci(cypress): split test suite to avoid crashes in CI
